changed controller to fit other services format

+++ ms_community/src/main/java/com/example/ms_community/repository/CommunityRepository.java
@@ -3,4 +3,4 @@ package com.example.ms_community.repository;
 import org.springframework.data.repository.CrudRepository;
 import com.example.ms_community.model.Community;;
 
-public interface CommunityRepository extends CrudRepository<Community, Integer>{}
+public interface CommunityRepository extends CrudRepository<Community, Long>{}
